### 2021-02-06 #426 @amoeba
dataspice - Create lightweight schema.org descriptions of Data
Submitting Author: Bryce Mecum (@amoeba)
Other Authors: Carl Boetigger (@cboettig), Scott Chamberlaim (@sckott), Auriel Fournier (@aurielfournier), Kelly Hondula (@khondula), Anna Krystalli (@annakrystalli), Maëlle Salmon (@maelle), Kate Webbink (@magpiedin), Kara Woo (@karawoo)
Repository: https://github.com/ropenscilabs/dataspice/
Version submitted:  1.0.0
Version accepted: TBD

---

```
Package: dataspice
Version: 1.0.0
Title: Create Lightweight Schema.org Descriptions of Data
Description: The goal of 'dataspice' is to make it easier for researchers to
  create basic, lightweight, and concise metadata files for their datasets.
  These basic files can then be used to make useful information available during
  analysis, create a helpful dataset "README" webpage, and produce more complex
  metadata formats to aid dataset discovery. Metadata fields are based on
  the 'Schema.org' and 'Ecological Metadata Language' standards.
    person("Carl", "Boettiger", role = c("aut"), comment = "https://github.com/cboettig"),
    person("Scott", "Chamberlain", role = c("aut"), comment = "https://github.com/sckott"),
    person("Auriel", "Fournier", role = c("aut"), comment = "https://github.com/aurielfournier"),
    person("Kelly", "Hondula", role = c("aut"), comment = "https://github.com/khondula"),
    person("Anna", "Krystalli", role = c("aut"), comment = "https://github.com/annakrystalli"),
    person("Bryce", "Mecum", role = c("aut", "cre"), email = "brycemecum@gmail.com", comment = "https://github.com/amoeba"),
    person("Maëlle", "Salmon", role = c("aut"), comment = "https://github.com/maelle"),
    person("Kate", "Webbink", role = c("aut"), comment = "https://github.com/magpiedin"),
    person("Kara", "Woo", role = c("aut"), comment = "https://github.com/karawoo"),
    person("Irene", "Steves", role = c("ctb"), comment = "https://github.com/isteves"))
    EML,
    fs,
    whisker,
    tools,
    shiny,
    rhandsontable,
    magrittr
    kableExtra,
    servr,
    listviewer,
    maps
```

    `dataspice` helps create metadata for deposit/publication of data

    The target audience is people producing data of all types, scientists and otherwise. Documenting data has a multitude of scientific applications including data publication, data sharing, data integration, etc.

    No other packages do the same thing (to my knowledge). `dataspice` is somewhat related to [EML](https://docs.ropensci.org/EML/), similar to [codebook](https://cran.r-project.org/web/packages/codebook/index.html) and [EMLAssemblyLine](https://github.com/EDIorg/EMLassemblyline).

    I don't think this is applicable to our package.

    @maelle and I have talked about this on ropensci Slack
