### 2021-02-19 #426 @emilyriederer
### Editor checks:

---

#### Editor comments

Hi Team! Thank you so much for your submission. I have run the preliminary checks and overall everything looks good. In summary, below I highlight a few small spelling mistakes, one styling issue flagged by `goodpractice`, and one question about the current testing plan. Please let me know what you think.

##### Spelling

The following may be spelling mistakes. I see "shinyapp" and "Shinyapps" are consistently used as single words, though, so I am open to that being a style choice.

```
> spelling::spell_check_package()
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD                    FOUND IN
rectagular              populating-templates.Rmd:22
shinyapp                edit_access.Rd:14
                        edit_attributes.Rd:14
                        edit_biblio.Rd:14
                        edit_creators.Rd:14
Shinyapps               edit_access.Rd:5
                        edit_attributes.Rd:5
                        edit_biblio.Rd:5
                        edit_creators.Rd:5
worfklowdiagram         README.md:54
                        README.Rmd:56
```

##### Styling

`goodpractice` highlights many cases where lines longer than the recommended 80 characters:

```
> goodpractice::gp()
```

- R/build_site.R
- R/edit_file.R
- R/eml_crosswalk.R
- R/eml_to_spice.R
- R/jsonld_to_mustache.R
- R/prep.R
- R/prep_access.R
- R/serve_site.R
- R/spice_to_eml.R
- R/validate_metadata.R
- R/write_spice.R

- tests/testthat/setup.R
- tests/testthat/test-crosswalks.R
- tests/testthat/test-eml_to_spice.R
- tests/testthat/test-prep_attribute.R
- tests/testthat/test-write_spice.R

##### Continuous integration

I see your testing R CMD CHECK across multiple OS on GitHub Actions. Do you plan to run your `testthat` unit tests here as well? The don't seem to be integrated currently.

---

Reviewers:
Due date:
