### 2021-02-22 115cf38
Update README to point at CRAN install
