+++ README.md
@@ -1,7 +1,6 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-dataspice
-=========
+# dataspice
 
 ![CRAN Version](https://www.r-pkg.org/badges/version/dataspice)
 ![CI](https://github.com/ropenscilabs/dataspice/workflows/R-CMD-check/badge.svg)
@@ -25,8 +24,7 @@ standards](https://github.com/ropenscilabs/dataspice#resources) and
 represent a lowest common denominator which means converting between
 formats should be relatively straightforward.
 
-Example
--------
+## Example
 
 A fully worked example can be found
 [here](https://github.com/amoeba/dataspice-example) and a live preview
@@ -34,22 +32,23 @@ of the output [here](https://amoeba.github.io/dataspice-example/). An
 example of how Google sees this can be found
 [here](https://search.google.com/structured-data/testing-tool/u/0/#url=https%3A%2F%2Famoeba.github.io%2Fdataspice-example%2F).
 
-Installation
-------------
+## Installation
 
-You can install the development version from
-[GitHub](https://github.com/) with:
+You can install the latest version from
+[CRAN](https://cran.r-project.org):
 
-    # install.packages("devtools")
-    devtools::install_github("ropenscilabs/dataspice")
+``` r
+install.packages("dataspice")
+```
 
-Workflow
---------
+## Workflow
 
-    create_spice()
-    # Then fill in template CSV files, more on this below
-    write_spice()
-    build_site() # Optional
+``` r
+create_spice()
+# Then fill in template CSV files, more on this below
+write_spice()
+build_site() # Optional
+```
 
 ![worfklowdiagram](man/figures/dataspice_workflow.png)
 
@@ -87,21 +86,27 @@ the associated helper function and/or
 To see an example of how `prep_attributes()` works, load the data files
 that ship with the package:
 
-    data_files <- list.files(system.file("example-dataset/", package = "dataspice"),
-                             pattern = ".csv",
-                             full.names = TRUE)
+``` r
+data_files <- list.files(system.file("example-dataset/", package = "dataspice"),
+                         pattern = ".csv",
+                         full.names = TRUE)
+```
 
 This function assumes that the metadata templates are in a folder called
 `metadata` within a `data` folder.
 
-    attributes_path <- file.path("data", "metadata", "attributes.csv")
+``` r
+attributes_path <- file.path("data", "metadata", "attributes.csv")
+```
 
 Using `purrr::map()`, this function can be applied over multiple files
 to populate the header names
 
-    data_files %>%
-      purrr::map(~ prep_attributes(.x, attributes_path),
-                 attributes_path = attributes_path)
+``` r
+data_files %>%
+  purrr::map(~ prep_attributes(.x, attributes_path),
+             attributes_path = attributes_path)
+```
 
 The output of `prep_attributes()` has the first two columns filled out:
 
@@ -260,10 +265,10 @@ temporal coverage
 
 `creators.csv` has one row for each of the dataset authors
 
-| id  | name           | affiliation                                           | email                                                                          |
-|:----|:---------------|:------------------------------------------------------|:-------------------------------------------------------------------------------|
-| NA  | Jeanette Clark | National Center for Ecological Analysis and Synthesis | <a href="mailto:jclark@nceas.ucsb.edu" class="email">jclark@nceas.ucsb.edu</a> |
-| NA  | Rich,Brenner   | Alaska Department of Fish and Game                    | richard.brenner.alaska.gov                                                     |
+| id  | name           | affiliation                                           | email                      |
[…]
