### 2021-02-22 063424e
Address spelling issues from onboarding review
- Use "Shiny" and "Shiny app" consistently
- Fix a typo "rectagular"
- Add a proper alt text description for the workflow diagram image
