+++ vignettes/populating-templates.Rmd
@@ -19,11 +19,11 @@ The `dataspice` metadata fields are based on [Schema.org](https://schema.org/) a
 
 ## Step 1 - Start with one or more data files
 
-To start the user will have one or more datafiles in a common directory. We currently support rectagular data with headers in a .csv file or spatial data with attributes. 
+To start the user will have one or more datafiles in a common directory. We currently support rectangular data with headers in a .csv file or spatial data with attributes.
 
 ## Step 2 - Fill in Templates
 
-`create_spice()` reads in the files from that directory and creates a set of template metadata files for the user to populate. These metadata templates will include some data extracted from the user's datafiles, including things like file names and measured variable names to aid the user in populating the metadata files. 
+`create_spice()` reads in the files from that directory and creates a set of template metadata files for the user to populate. These metadata templates will include some data extracted from the user's datafiles, including things like file names and measured variable names to aid the user in populating the metadata files.
 
 Once these are created, the user needs to open each template, fill in the missing metadata (as completely as possible), and read the files back into R.
 
@@ -33,21 +33,19 @@ Once these are created, the user needs to open each template, fill in the missin
 - `attributes.csv`: This is where most of the user data entry will take place. For each variable, its name, units, and a written description are filled in.
 - `biblio.csv`: Citation information about the project, as much or as little data as possible can be included, but if things like bounding box coordinates are not included, then when the website is generated there will not be a bounding box map generated.
 - `access.csv`: Includes a row for each file that was read in, and documents the name of each file and its format.
-  
+
 ![](https://user-images.githubusercontent.com/6106733/40335756-444642e0-5d1a-11e8-8c33-0d3053441a79.png)
 
-## Step 3 - Save metadata in JSON 
+## Step 3 - Save metadata in JSON
 
-Once the populated the metadata files are read back R they can then be fed into `write_JSON()`, which converts those tabular metadata files into JSON files. 
+Once the populated the metadata files are read back R they can then be fed into `write_JSON()`, which converts those tabular metadata files into JSON files.
 
 ## Step 4 - Create a dataspice website
 
-A dataset README website is an interactive representation of the JSON information about the data. Assuming sufficient information is provided in the JSON it will include a map of the points and a bounding box of the area of study. 
+A dataset README website is an interactive representation of the JSON information about the data. Assuming sufficient information is provided in the JSON it will include a map of the points and a bounding box of the area of study.
 
 The output from `write_JSON()` is fed into `build_site()` to create the website.
 
 `serve_site()` ?
 
 ![dataspice-website](../man/figures/website_example.png)
-
-
