+++ README.Rmd
@@ -52,7 +52,7 @@ write_spice()
 build_site() # Optional
 ```
 
-![worfklowdiagram](man/figures/dataspice_workflow.png)
+![diagram showing a workflow for using dataspice](man/figures/dataspice_workflow.png)
 
 ### Create spice
 
@@ -112,12 +112,12 @@ readr::read_csv(
 
 Each of the metadata templates can be edited interactively using a [Shiny](https://shiny.rstudio.com/) app:
 
-* `edit_attributes()` opens a Shiny app that can be used to edit `attributes.csv`. The shiny app displays the current `attributes` table and lets the user fill in an informative description and units (e.g. meters, hectares, etc.) for each variable.
+* `edit_attributes()` opens a Shiny app that can be used to edit `attributes.csv`. The Shiny app displays the current `attributes` table and lets the user fill in an informative description and units (e.g. meters, hectares, etc.) for each variable.
 * `edit_access()` opens an editable version of `access.csv`
 * `edit_creators()` opens an editable version of `creators.csv`
 * `edit_biblio()` opens an editable version of `biblio.csv`
 
-![edit_attributes shiny app](man/figures/edit_attributes.png)
+![edit_attributes Shiny app](man/figures/edit_attributes.png)
 
 Remember to click on **Save** when finished editing.
 
