### 2021-02-22 82b5d9e
Fix majority of goodpractice violations
