+++ R/eml_crosswalk.R
@@ -247,7 +247,8 @@ crosswalk <- function(doc, term) {
 #' # Now fill in the attributes data.frame. See `EML::set_attributes`.
 #'
 #' # And last, set the attributes on our EML document
-#' eml_doc$dataset$dataTable[[1]]$attributeList <- EML::set_attributes(attributes)
+#' eml_doc$dataset$dataTable[[1]]$attributeList <-
+#'   EML::set_attributes(attributes)
 #' }
 crosswalk_variables <- function(spice) {
   if (is.character(spice)) {
