### 2021-02-22 1adb1a2
Add tests for build_site
