### 2021-02-22 66056d2
Run actions on any push or PR
