### 2021-02-22 #426 @amoeba
Hey again @emilyriederer: I've had a chance to go over your review and didn't run into any problems. I'll address your feedback in the same structure you provided it:

- **Spelling**: All fixed. I went with "Shiny app" everywhere as that seems to be the preferred nomenclature.
- **Styling**: I mostly was able to address these but stopped short of editing some of the lines that crossed the 80 character mark (though I fixed many). The ones I didn't change seemed to me like they'd become less readable if I did any reformatting, e.g., https://github.com/ropenscilabs/dataspice/blob/1adb1a25cf19b5caee8ecd4f5e1dc3971e9a2a48/R/edit_file.R#L94-L96. Please let me know you think this is acceptable.
- **Continuous integration**: It's my understanding that the GH Action we run _does_ run the tests when it runs R CMD CHECK. I double-checked by checking the logs of the [latest run](https://github.com/ropenscilabs/dataspice/runs/1948560664?check_suite_focus=true) and I see:
    ```
    * checking tests ...
      Running ‘testthat.R’
      OK
    ```

    Do you agree with my take?

Thanks again for the review -- the changes solidly tightened up some parts of the package. Let me know if you have any questions. My changes are pushed to the new main branch (was master) and can be seen at https://github.com/ropenscilabs/dataspice/tree/1adb1a25cf19b5caee8ecd4f5e1dc3971e9a2a48.
