### 2021-02-24 #426 @emilyriederer
Thanks for the replies, @amoeba ! That all makes sense and looks good to me. My apologies on the last issue; I checked the log for the CI but somehow missed where the tests were getting called. I'm on board now. Have you all considered using `covr` to show test coverage statistics in the repo also?

I've begun the search for reviewers and have confirmed @aebratt as the first reviewer. I will hopefully confirm the second shortly.
