### 2021-02-24 #426 @amoeba
> Have you all considered using covr to show test coverage statistics in the repo also?

Not as a group but I had thought about it in light of onboarding and [Continuous Integration Best Practices](https://devguide.ropensci.org/ci.html). Might be good to set up a separate test workflow and loop in covr. If we go that route, should I aim to have that working as part of onboarding or could it be done after?
