### 2021-02-24 ropenscilabs/dataspice#97 @amoeba
Add Test and covr workflows
This was triggered by [onboarding](https://github.com/ropensci/software-review/issues/426) and is recommended in [Continuous Integration Best Practices](https://devguide.ropensci.org/ci.html).
