+++ README.md
@@ -4,6 +4,8 @@
 
 ![CRAN Version](https://www.r-pkg.org/badges/version/dataspice)
 ![CI](https://github.com/ropenscilabs/dataspice/workflows/R-CMD-check/badge.svg)
+[![Codecov test
+coverage](https://codecov.io/gh/ropenscilabs/dataspice/branch/main/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/dataspice?branch=main)
 
 The goal of `dataspice` is to make it easier for researchers to create
 basic, lightweight, and concise metadata files for their datasets by
