### 2021-02-24 ropenscilabs/dataspice#97 @amoeba
Decided not to have a separate action just for tests (in order to show a "tests passing" badge) since no one seems to do that. I did add a workflow for covr in 63d3a238ced02042288c589e6e61212e8c6597a0 and updated the readme with a badge.
