### 2021-03-12 #426 @aebratt
- NA

  - Looks great. I think this is a compelling package with a lot of use cases.
  - Installed well on my machine.
  - Vignette runs locally and demonstrates some functionality. However, it does not include
  a description of the `edit_xx` functions which open a Shiny app for editing metadata attributes.
  In my opinion this is a nice piece of functionality in this package.
  - Looks good.
- [x] **Examples** (that run successfully locally) for all exported functions
  - This is a nit-pick but I think that calling the example [here](https://github.com/amoeba/dataspice-example) fully-worked is somewhat misleading, since the csv editing is happening outside of R, for most. I think that could be made clearer.
  - I also think that [this](https://annakrystalli.me/dataspice-tutorial/) tutorial by one of the authors is excellent, but it's not clearly linked from other documentation.
  - Looks good.

  - Installed well.
  - I explored the package with a few of my own use-cases without any issues.
  - Again, I explored the package with a few of my own use-cases without any issues.
  - Was able to pass tests but I agree with the above discussion that using `covr` to show test coverage statistics within the repository would be nice to have. Has there been progress made on that front?
  - Looks good.

Estimated hours spent reviewing: 4 hours, much of which was playing with use cases.

---

This is a really slick package that is well-done and widely applicable. Nice work, team! My primary suggestion would be to standardize the documentation so that the pacakage is easy to learn from multiple points of entry (see suggestions about README, Example, Vignette, and Tutorial above). In this documentation I would like to see the Shiny functionality reiterated. I am also in support of using `covr`. Additionally, I noted a couple of small style notes below:

#### Style

I am on board with keeping the few remaining long lines long for stylistic reasons. I did notice
what I think is a typo:
```
spatialCoverge          parse_GeoShape_box.Rd:5,16
                        parse_spatialCoverage.Rd:5,16
```
and I am guessing that this should say "spatialCoverage". It is just in the documentation though.

I also noticed that the pacakge imports `spread` from `tidyR` which I believe was deprecated by `pivod_wider`. Is that intentional?
