### 2021-03-21 #426 @tdjames1
- Scott Chamberlain: Co-member of sAPROPOS working group, Feb 2017
  - Anna Krystalli: Colleague and co-organiser of Sheffield R Users Group

- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work.

- [ ] **Examples** (that run successfully locally) for all exported functions

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This is a neat, self-contained package that makes it easy for dataset owners to create standardised metadata to improve visibility and usability of their data.

#### Package installation and checks

Installation and package checks ran smoothly except size reduction checks which produced a warning about `qpdf` - I do have this installed so I'm not sure why these checks were not carried out.

`goodpractice` reported several problems with long lines, relating to (1) strings providing help text for the Shiny app and (2) test invocations of the `build_site()` function.  For (1), you could use the flexibility in the `...` argument to the Shiny HTML builder functions which results in unnamed arguments becoming children of the element, e.g.:

```
shiny::h6("fileName: The name of the input data file(s). ",
          "Don't change this.")
```

For (2), you could create an intermediate variable to hold the file path or break the line before the package argument.

`goodpractice` also flagged up calls to `setwd()` in test functions. However, the authors have already followed the advice given by `goodpractice` about "using on.exit() to restore the working directory."

#### Package metadata

- `devtools::spell_check()` complains about missing Language field in `DESCRIPTION`.
- Suggest fixing misspelling of `spatialCoverage` in documentation for  `parse_GeoShape_box` and `parse_spatialCoverage` functions.

**README**

- In online documentation (https://docs.ropensci.org/dataspice), link to "metadata standards" opens up https://github.com/ropenscilabs/dataspice#resources rather than linking to https://docs.ropensci.org/dataspice/#resources.

- "An example of how Google sees this can be found here." leads to a page which contains a banner stating that the structured data testing tool is being deprecated. Can you find an alternative demonstration of validating the structured data?

- _Convert to EML_ section: the output produced by `spice_to_eml(spice)` is very wide and not easy to read, is there any way to resolve this via knitr options? e.g. https://bookdown.org/yihui/rmarkdown-cookbook/text-width.html
There is also a missing space following a full stop in the output: "produced.You".

**Help files**
- No top-level documentation `?dataspice`. Consider `usethis::use_package_doc()` to add a template for the top-level documentation.

- `validate_file_path()` help text is headed "Prepare attributes" presumably due to shared content - it's a bit confusing. Also no example usage for this function in the example.

- `prepare_access()` - no example given.

- `eml_to_spice()` - typo in example "disd".

- `write_spice()` - no example given.

**Shiny app**

- 'The units the variable was measured in.' -> 'The units in which the variable was measured.'

**Vignettes**

_Creating Schema.org Metadata and README website with dataspice_
- Set width of figures to match page width.
- Under Step 4, final sentence appears to be incomplete: "serve_site() ?"

Worked well with provided datasets and other trial datasets. A couple of comments on the Shiny app user interface:

`edit_attributes()`: help text for `fileName`, `varableName` is marked 'Do not change' - is it possible to enforce this by setting the columns to read only?

`edit_biblio()`: the list of metadata covered is rather long, which makes it hard to read the help text while filling in the table. Some possible solutions: offer help text as onhover or onclick of table headings; split up table into chunks to match the metadata headings.

Source code is well-formulated, clear and free from duplication.

Some file naming could be adjusted to match function naming e.g. `prep.R` -> `prep_attributes.R`.

Test coverage is a bit low and could be improved by providing some missing tests e.g. `prep_access()`, `validate_metadata()`.

Also, rOpenSci Packaging guide suggests testing shiny app components: "Packages with shiny apps should use a unit-testing framework such as `shinytest` to test that interactive interfaces behave as expected."
