### 2021-03-22 #426 @emilyriederer
Thank you @aebratt and @tdjames1 for the detailed reviews and helpful feedback! @amoeba , the ball is back in your court now! In the next few weeks, please take some time to consider the feedback, engage with the reviewers, and make any relevant changes.

In the meantime, I have a few more administrative favors to ask of you:

1. Please add a badge to your `README` denoting that the package is under rOpenSci review. You may do this with the `rodev::use_review_badge(426)` command or by directly pasting the following in your MD file:

2. Please also add a `NEWS.md` document to your repo also

Thanks again to all!
