+++ NEWS.md
@@ -1,0 +1,6 @@
+# NEWS
+
+## dataspice 1.0.0
+
+Initial CRAN release
+
