### 2021-04-12 #426 @amoeba
Hey again @emilyriederer, @aebratt, and @tdjames1. I'm a bit late in getting to all of your feedback but have carved out some time tonight to start. So far I've collected all of the feedback in order to see how much time it might take me and I think I can get everything addressed within the week. Apologies for the delay!
