+++ tests/testthat/test-edit_file.R
@@ -1,0 +1,5 @@
+context("edit_file")
+
+test_that("edit_file fails when it can't find anything", {
+  expect_error(edit_file(tempdir()))
+})
