+++ man/dataspice-package.Rd
@@ -1,0 +1,45 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dataspice-package.R
+\docType{package}
+\name{dataspice-package}
+\alias{dataspice}
+\alias{dataspice-package}
+\title{dataspice: Create Lightweight Schema.org Descriptions of Data}
+\description{
+The goal of 'dataspice' is to make it easier for researchers to
+  create basic, lightweight, and concise metadata files for their datasets.
+  These basic files can then be used to make useful information available during
+  analysis, create a helpful dataset "README" webpage, and produce more complex
+  metadata formats to aid dataset discovery. Metadata fields are based on
+  the 'Schema.org' and 'Ecological Metadata Language' standards.
+}
+\seealso{
+Useful links:
+\itemize{
+  \item \url{https://github.com/ropenscilabs/dataspice}
+  \item Report bugs at \url{https://github.com/ropenscilabs/dataspice/issues}
+}
+
+}
+\author{
+\strong{Maintainer}: Bryce Mecum \email{brycemecum@gmail.com} (https://github.com/amoeba)
+
+Authors:
+\itemize{
+  \item Carl Boettiger (https://github.com/cboettig)
+  \item Scott Chamberlain (https://github.com/sckott)
+  \item Auriel Fournier (https://github.com/aurielfournier)
+  \item Kelly Hondula (https://github.com/khondula)
+  \item Anna Krystalli (https://github.com/annakrystalli)
+  \item Maëlle Salmon (https://github.com/maelle)
+  \item Kate Webbink (https://github.com/magpiedin)
+  \item Kara Woo (https://github.com/karawoo)
+}
+
+Other contributors:
+\itemize{
+  \item Irene Steves (https://github.com/isteves) [contributor]
+}
+
+}
+\keyword{internal}
