### 2021-04-12 e820f82
Add a space to the message in spice_to_eml
