### 2021-04-12 cb84373
Improve the Example section of the README
