+++ man/validate_file_paths.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/prep.R
 \name{validate_file_paths}
 \alias{validate_file_paths}
-\title{Prepare attributes}
+\title{Validate file paths}
 \usage{
 validate_file_paths(data_path = "data", ...)
 }
@@ -21,32 +21,17 @@ attributes returned using \code{names()} can be extracted from r object, stored 
 filter files for patterns.}
 }
 \value{
-\code{prep_attributes()} updates the \code{attributes.csv} and writes to
-\code{attributes_path}.
+One or more data file paths
 }
 \description{
-Extract \code{variableNames} from data file(s) and add them to \code{attributes.csv}.
-The helper \code{\link{validate_file_paths}} can be used to create vectors of
-valid file paths that can be checked and then passed as \code{data_path} argument
-to \code{\link{prep_attributes}}.
+Helper function to return a set of file paths for use in other functions
 }
 \examples{
 \dontrun{
-create_spice()
-# extract attributes from all `csv`, `tsv`, `rds` files in the data folder
-# (non recursive)
-prep_attributes()
-# recursive
-prep_attributes(recursive = TRUE)
-# extract attributes from a single file using file path
-data_path <- system.file("example-dataset","BroodTables.csv",
-                         package = "dataspice")
-prep_attributes(data_path)
-# extract attributes from a single file by file path pattern matching
-data_path <- system.file("example-dataset", package = "dataspice")
-prep_attributes(data_path, pattern = "StockInfo")
-# extract from a folder using folder path
-data_path <- system.file("example-dataset", package = "dataspice")
-prep_attributes(data_path)
+# Assuming some data files in "./data"
+my_files <- validate_file_paths()
+
+# If your data files are in `another_folder`
+my_files <- validate_file_paths("another_folder")
 }
 }
