### 2021-04-12 dec3dcf
Fix bug I introduced in edit_file
