### 2021-04-29 d9ea268
Finish last set of changes from ropensci review
https://github.com/ropensci/software-review/issues/426
