+++ R/edit_file.R
@@ -63,7 +63,8 @@ edit_file <- function(metadata_dir = "data/metadata", file) {
       shiny::titlePanel(paste0("Populate the ", file, " table")),
       shiny::helpText("Interactive application to read in ",
                       shiny::code("dataspice"),
-                      "metadata templates and populate with your own metadata."),
+                      "metadata templates and populate with your own ",
+                      "metadata."),
 
       # Header panel
       shiny::wellPanel(
@@ -82,49 +83,120 @@ edit_file <- function(metadata_dir = "data/metadata", file) {
           if (file == "attributes.csv") {
             list(
               shiny::h4("Variable metadata"),
-              shiny::h6("fileName: The name of the input data file(s). Don't change this."),
-              shiny::h6("variableName: The name of the measured variable. Don't change this."),
-              shiny::h6("description: A written description of the variable measured."),
-              shiny::h6("unitText: The units in which the variable was measured."),
-              shiny::helpText("Use", shiny::code("prep_attributes()"), "to extract", shiny::strong("variableName"), " from data files.")
+              shiny::h6(
+                "fileName: The name of the input data file(s). ",
+                "Don't change this."
+              ),
+              shiny::h6(
+                "variableName: The name of the measured variable. ",
+                "Don't change this."
+              ),
+              shiny::h6(
+                "description: A written description of the variable measured."
+              ),
+              shiny::h6(
+                "unitText: The units in which the variable was measured."
+              ),
+              shiny::helpText(
+                "Use",
+                shiny::code("prep_attributes()"),
+                "to extract",
+                shiny::strong("variableName"),
+                " from data files.")
             )
           },
           if (file == "biblio.csv") {
             list(
               shiny::h4("Bibliographic metadata"),
               shiny::h6("title: Title of the dataset(s) described."),
-              shiny::h6("description: Description of the dataset(s) described (aka abstract)."),
-              shiny::h6('datePublished: The date published in ISO 8601 format (YYYY-MM-DD).'),
-              shiny::h6("citation: Citation to a related work, such as a reference publication."),
-              shiny::h6("keywords: Keywords, separated by commas, associated with the dataset(s) described."),
-              shiny::h6("license: License under which (meta)data are published."),
-              shiny::h6("funder: Name of funders associated with the work through which data where generated."),
+              shiny::h6(
+                "description: Description of the dataset(s) ",
+                "described (aka abstract)."
+              ),
+              shiny::h6(
+                "datePublished: The date published in ISO 8601 format ",
+                "(YYYY-MM-DD)."
+              ),
+              shiny::h6(
+                "citation: Citation to a related work, such as a reference ",
+                "publication."
+              ),
+              shiny::h6(
+                "keywords: Keywords, separated by commas, associated with the ",
+                "dataset(s) described."
+              ),
+              shiny::h6(
+                "license: License under which (meta)data are published."
+              ),
+              shiny::h6(
+                "funder: Name of funders associated with the work through ",
+                "which data where generated."
+              ),
 
               shiny::br(),
-              shiny::h4("Spatial Coverage metadata"),
-              shiny::h6("geographicDescription: Description of the area of study. i.e., Crater Lake."),
-              shiny::h6("northBoundCoord: Southern latitudinal boundary of the data coverage area. For example 37.42242."),
-              shiny::h6("eastBoundCoord: Eastern longitudinal boundary of the data coverage area. For example -122.08585."),
-              shiny::h6("southBoundCoord: Northern latitudinal boundary of the data coverage area."),
-              shiny::h6("westBoundCoord: Western longitudinal boundary of the data coverage area."),
[…]
