+++ man/validate_file_paths.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/prep.R
+% Please edit documentation in R/prep_attributes.R
 \name{validate_file_paths}
 \alias{validate_file_paths}
 \title{Validate file paths}

[…]
