+++ README.md
@@ -1,11 +1,3 @@
----
-output:
-  md_document:
-    variant: gfm
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
 # dataspice
 
 ![CRAN Version](https://www.r-pkg.org/badges/version/dataspice)
@@ -20,12 +12,12 @@ editing the kind of files they’re probably most familiar with: CSVs. To
 spice up their data with a dash of metadata. These metadata files can
 then be used to:
 
--   Make useful information available during analysis.
--   Create a helpful dataset README webpage for your data similar to how
-    [pkgdown](https://pkgdown.r-lib.org/) creates websites for R
-    packages.
--   Produce more complex metadata formats for richer description of your
-    datasets and to aid dataset discovery.
+- Make useful information available during analysis.
+- Create a helpful dataset README webpage for your data similar to how
+  [pkgdown](https://pkgdown.r-lib.org/) creates websites for R
+  packages.
+- Produce more complex metadata formats for richer description of your
+  datasets and to aid dataset discovery.
 
 Metadata fields are based on
 [Schema.org/Dataset](https://schema.org/Dataset) and other [metadata
@@ -51,13 +43,13 @@ repo](https://github.com/annakrystalli/dataspice-tutorial)).
 You can install the latest version from
 [CRAN](https://cran.r-project.org):
 
-``` r
+```r
 install.packages("dataspice")
 ```
 
 ## Workflow
 
-``` r
+```r
 create_spice()
 # Then fill in template CSV files, more on this below
 write_spice()
@@ -74,12 +66,12 @@ default created in the `data` folder in the current working directory).
 
 The template files are:
 
--   **biblio.csv** - for title, abstract, spatial and temporal coverage,
-    etc.
--   **creators.csv** - for data authors
--   **attributes.csv** - explains each of the variables in the dataset
--   **access.csv** - for files, file types, and download URLs (if
-    appropriate)
+- **biblio.csv** - for title, abstract, spatial and temporal coverage,
+  etc.
+- **creators.csv** - for data authors
+- **attributes.csv** - explains each of the variables in the dataset
+- **access.csv** - for files, file types, and download URLs (if
+  appropriate)
 
 ### Fill in templates
 
@@ -90,18 +82,18 @@ the associated helper function and/or
 
 #### Helper functions
 
--   `prep_attributes()` populates the **`fileName`** and
-    **`variableName`** columns of the `attributes.csv` file using the
-    header row of the data files.
+- `prep_attributes()` populates the **`fileName`** and
+  **`variableName`** columns of the `attributes.csv` file using the
+  header row of the data files.
 
--   `prep_access()` populates the **`fileName`**, **`name`** and
-    **`encodingFormat`** columns of the `access.csv` file from the files
-    in the folder containing the data.
+- `prep_access()` populates the **`fileName`**, **`name`** and
+  **`encodingFormat`** columns of the `access.csv` file from the files
+  in the folder containing the data.
 
 To see an example of how `prep_attributes()` works, load the data files
 that ship with the package:
 
-``` r
+```r
 data_files <- list.files(system.file("example-dataset/", package = "dataspice"),
   pattern = ".csv",
   full.names = TRUE
@@ -111,14 +103,14 @@ data_files <- list.files(system.file("example-dataset/", package = "dataspice"),
 This function assumes that the metadata templates are in a folder called
 `metadata` within a `data` folder.
 
-``` r
+```r
 attributes_path <- file.path("data", "metadata", "attributes.csv")
 ```
 
 Using `purrr::map()`, this function can be applied over multiple files
 to populate the header names
 
-``` r
+```r
 data_files %>%
   purrr::map(~ prep_attributes(.x, attributes_path),
[…]
