### 2021-04-29 ropenscilabs/dataspice#101 @amoeba
Migrate from spread to pivot_wider
`tidyr::spread` is used in the package but it's superseded. See https://tidyr.tidyverse.org/reference/spread.html.
