### 2021-04-29 ropenscilabs/dataspice#102 @amoeba
Address width and scrolling in Shiny apps
`edit_biblio`, for example, gets really wide and unwieldy. We might use another approach there or at least find a way to keep the help documentation visible while you scroll.
