### 2021-04-29 #426 @amoeba
Hey @emilyriederer, apologies for the delay. I've gone over both reviews and made changes and am ready to go over my changes with the reviewers.

Thanks @aebratt and @tdjames1 for the very thorough dives into the package and all of the suggestions. Just about every suggestion or idea was reasonable and I've made changes in a branch at https://github.com/ropenscilabs/dataspice/tree/onboarding-comments.

A full item-by-item response to all of your comments is at the bottom under "Full comments" but they're mostly me agreeing with you and making the change. I've included it in the hope that it makes it easier for you both to verify I've addressed your comments.

For all of your comments that required a conversation, I've put them directly below under individual headers. Please have a look and let me know what you think.

### Code Coverage / shinytest (@emilyriederer, @aebratt , @tdjames1 )

This has come up a few times during review and I've done my best to bump up coverage. The package is now at ~72%

```
> covr::package_coverage()
dataspice Coverage: 72.17%
R/serve_site.R: 0.00%
R/edit_file.R: 0.47%
R/prep_attributes.R: 91.89%
R/prep_access.R: 94.74%
R/eml_to_spice.R: 98.08%
R/write_spice.R: 98.28%
R/eml_crosswalk.R: 98.57%
R/build_site.R: 100.00%
R/create_spice.R: 100.00%
R/jsonld_to_mustache.R: 100.00%
R/Person.R: 100.00%
R/PropertyValue.R: 100.00%
R/spice_to_eml.R: 100.00%
R/validate_metadata.R: 100.00%
R/write_jsonld.R: 100.00%
```

`edit_file.R` (mostly) and `serve_site.R` are dragging the average down. `edit_file.R` would require some work to get high coverage. @emilyriederer: Is this is an acceptable level of coverage given the nature of the code?

### use of `tidyr::spread` (@aebratt)

I think we used this just because the pivot_* functions weren't around. `spread` is marked as `superceded`. I've filed an issue for the future at https://github.com/ropenscilabs/dataspice/issues/101 but I'd like to keep it as is for now. Sound good?

### spice_to_eml output width (@tdjames1)

I wasn't able to find anything that quite did the trick so I'm inclined to leave it as is. Is that alright?

### edit_biblio() width (@tdjames)

> edit_biblio(): the list of metadata covered is rather long, which makes it hard to read the help text while filling in the table. Some possible solutions: offer help text as onhover or onclick of table headings; split up table into chunks to match the metadata headings.

I'm inclined to leave it as is but have [filed an issue](https://github.com/ropenscilabs/dataspice/issues/102) to track it. Sound good?

---

Hopefully I've addressed all of your comments @aebratt and @tdjames1. Let me know if that's not the case. I'm looking forward to hearing what you both think and thanks again for the thorough reviews. This has been super helpful so far.

<summary>Full comments</summary>

## @aebratt review

> Vignette runs locally and demonstrates some functionality. However, it does not include
a description of the edit_xx functions which open a Shiny app for editing metadata attributes.
In my opinion this is a nice piece of functionality in this package.

Good catch. I think this vignette was written before the Shiny apps were made and we didn't update the vignette. Fixed.

> This is a nit-pick but I think that calling the example here fully-worked is somewhat misleading, since the csv editing is happening outside of R, for most. I think that could be made clearer.

I agree with your take so I've updated that section to be more clear about what is being shown.
We use the phrasing "An basic example repository" now instead.

> I also think that this tutorial by one of the authors is excellent, but it's not clearly linked from other documentation.

Right, I'm not sure I had seen that and I've linked it in the Example section.

> Was able to pass tests but I agree with the above discussion that using covr to show test coverage statistics within the repository would be nice to have. Has there been progress made on that front?

I've gone through and bumped up test coverage for the package as a whole to ~80% and all files that really can be tested to over 90%.

> In this documentation I would like to see the Shiny functionality reiterated. I am also in support of using covr.

> Additionally, I noted a couple of small style notes below:

> spatialCoverge typo

Fixed.

> I also noticed that the pacakge imports spread from tidyR which I believe was deprecated by pivod_wider. Is that intentional?

Just timing. I'm not sure the pivot_* were around then? I'll file an issue to get converted over post-review. `spread` isn't deprecated but ["superceded"](https://lifecycle.r-lib.org/articles/stages.html#superseded).

## @tdjames1' review:

> goodpractice reported several problems with long lines

I've fixed all of these.

> devtools::spell_check() complains about missing Language field in DESCRIPTION.

Fixed.

> Suggest fixing misspelling of spatialCoverage in documentation for parse_GeoShape_box and parse_spatialCoverage functions.

Fixed.

> In online documentation (https://docs.ropensci.org/dataspice), link to "metadata standards" opens up https://github.com/ropenscilabs/dataspice#resources rather than linking to https://docs.ropensci.org/dataspice/#resources.

Fixed with a relative link so the link works everywhere.

> "An example of how Google sees this can be found here." leads to a page which contains a banner stating that the structured data testing tool is being deprecated. Can you find an alternative demonstration of validating the structured data?

Fixed. Google has a new tool and I've updated the link.

> Convert to EML section: the output produced by spice_to_eml(spice) is very wide and not easy to read, is there any way to resolve this via knitr options? e.g. https://bookdown.org/yihui/rmarkdown-cookbook/text-width.html

I tried `options(width...)` and a few other things and wasn't able to come up with something. I'm inclined to leave things as is since the messages appear fine in R. What do you think?

> There is also a missing space following a full stop in the output: "produced.You".

Fixed.

> No top-level documentation ?dataspice. Consider usethis::use_package_doc() to add a template for the top-level documentation.

Nice trick! Done and thanks!

> validate_file_path() help text is headed "Prepare attributes" presumably due to shared content - it's a bit confusing. > Also no example usage for this function in the example.

Fixed the incorrect title and set up a proper documentation page.

> prepare_access() - no example given.

Fixed.

> eml_to_spice() - typo in example "disd".

Fixed.

> write_spice() - no example given.

Fixed.

> Shiny app: 'The units the variable was measured in.' -> 'The units in which the variable was measured.'

Fixed.

> Vignettes:
> Set width of figures to match page width.

Fixed.

> Under Step 4, final sentence appears to be incomplete: "serve_site() ?"

Fixed by removing it. That must have been a question the author had and never resolved.

> edit_attributes(): help text for fileName, varableName is marked 'Do not change' - is it possible to enforce this by setting the columns to read only?

This is a great idea. All I can find is a way to make these tables read-only in their entirety. I'll file an issue so we can look at this for the future.

> edit_biblio(): the list of metadata covered is rather long, which makes it hard to read the help text while filling in the table. Some possible solutions: offer help text as onhover or onclick of table headings; split up table into chunks to match the metadata headings.

Totally agree. This might take some deep digging so I'll file an issue for us to track it. Thanks.

> Code: Some file naming could be adjusted to match function naming e.g. prep.R -> prep_attributes.R.

Good point, changed.

> Test coverage is a bit low and could be improved by providing some missing tests e.g. prep_access(), validate_metadata().

I've gone through and bumped up test coverage for the package as a whole to ~80% and all files that really can be tested to over 90%.

> Also, rOpenSci Packaging guide suggests testing shiny app components: "Packages with shiny apps should use a unit-testing framework such as shinytest to test that interactive interfaces behave as expected."

Good catch. I knew about shinytest but hadn't looked at how to use it. From what I can tell, it requires the Shiny apps under test to be written as stand-alone apps which ours aren't. Refactoring would take a bit of time for sure. We might ask @emilyriederer to weigh in on whether this is a strict requirement or not.
