+++ vignettes/overview.Rmd
@@ -24,7 +24,7 @@ To start the user will have one or more datafiles in a common directory. We curr
 
 `create_spice()` reads in the files from that directory and creates a set of template metadata files for the user to populate. These metadata templates will include some data extracted from the user's datafiles, including things like file names and measured variable names to aid the user in populating the metadata files.
 
-Once these are created, the you need to fill in each template file with the remaining metadata.
+Once these are created, you need to fill in each template file with the remaining metadata.
 Templates can be filled either:
 
 1. With the integrated [Shiny](https://shiny.rstudio.com/) applications (`edit_attributes()`, `edit_access()`, `edit_creators()`, `edit_biblio()`)
