### 2021-05-06 #426 @tdjames1
## Reviewer Response

Hi @amoeba, thanks for the thorough response to my review. I'm satisfied that you've resolved the points that I raised. I've responded below to the specific queries that you flagged up.

I noticed a typo that was introduced in the overview vignette so I've made a pull request to fix that in your onboarding branch.

> ### Code Coverage / shinytest (@emilyriederer, @aebratt , @tdjames1 )
> This has come up a few times during review and I've done my best to bump up coverage. The package is now at ~72%

That seems pretty good to me.

> ### spice_to_eml output width (@tdjames1)
> I wasn't able to find anything that quite did the trick so I'm inclined to leave it as is. Is that alright?

Sure – it's only cosmetic so no problem to leave as is. Just one of those annoying things!

> ### edit_biblio() width (@tdjames)
> > edit_biblio(): the list of metadata covered is rather long, which makes it hard to read the help text while filling in the table. Some possible solutions: offer help text as onhover or onclick of table headings; split up table into chunks to match the metadata headings.
>
> I'm inclined to leave it as is but have [filed an issue](https://github.com/ropenscilabs/dataspice/issues/102) to track it. Sound good?

Sounds good.

> > edit_attributes(): help text for fileName, varableName is marked 'Do not change' - is it possible to enforce this by setting the columns to read only?
>
> This is a great idea. All I can find is a way to make these tables read-only in their entirety. I'll file an issue so we can look at this for the future.

This should be possible according to the [rhandsontable documentation](http://jrowen.github.io/rhandsontable/#read-only):
"The individual columns, rows and cells can also be set to readOnly"
I'm noting this point here as I can't find the issue.

#### Final approval (post-review)

Estimated hours spent reviewing: 7
