### 2021-05-11 ropenscilabs/dataspice#105 @amoeba
Make use of rhandsontable's read only features
Stemming from https://github.com/ropensci/software-review/issues/426, @tdjames1 mentions we might benefit from using rhandsontable's [read only](http://jrowen.github.io/rhandsontable/#read-only) functionality to prevent editing of certain columns.

This would be a good improvement to the Shiny apps and maybe even a good external contribution if someone was interested in taking this on.
