### 2021-05-11 ropenscilabs/dataspice#106 @amoeba
Add ropensci onboarding reviewers in package acknowledgement section
Add our two wonderful reviewers from https://github.com/ropensci/software-review/issues/426 to the readme and maybe even DESCRIPTION as their contributions were substantial.
