### 2021-05-14 #426 @aebratt
## Reviewer Response

Hi @amoeba. Apologies for the delay. Thanks for the thorough response to our reviews. I am quite happy with the changes you've made! In particular:

> ### Code Coverage / shinytest (@emilyriederer, @aebratt , @tdjames1 )
> This has come up a few times during review and I've done my best to bump up coverage. The package is now at ~72%

On board with this

> ### use of `tidyr::spread` (@aebratt)
> I think we used this just because the pivot_* functions weren't around. `spread` is marked as `superceded`. I've filed an issue for the future at [ropenscilabs/dataspice#101](https://github.com/ropenscilabs/dataspice/issues/101) but I'd like to keep it as is for now. Sound good?

Sounds good!

#### Final approval (post-review)

Estimated hours spent reviewing: 5
