### 2021-05-16 #426 @emilyriederer
Approved! Thanks @amoeba  for submitting and @aebratt @tdjames1  for your reviews! 😄

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with GitHub Actions)
- [ ] Please check you updated the package version post-review version updated and that you documented all changes in NEWS.md

We've put together an online book with our best practice and tips, [this chapter](https://devguide.ropensci.org/collaboration.html) starts the 3d section that's about guidance for after onboarding. Please tell us what could be improved, the corresponding repo is [here](https://github.com/ropensci/dev_guide). Additionally, since you plan to submit to CRAN, please check out our CRAN [gotchas](https://devguide.ropensci.org/building.html#crangotchas)
