### 2021-05-16 #426 @stefaniebutland
@amoeba If you're willing, I think this would make a great rOpenSci blog post for several reasons. Our community call [The Wild World of Data Repositories](https://ropensci.org/commcalls/dec2020-datarepos/) was our most popular ever. ~160 people from 19 countries on 5 continents attended and it was remarkable that in a poll of 133 attendees, 44% self-identified as being affiliated with a repository or infrastructure org; 56% not.

@sinarueeger wrote her perspective on the community call [The headache with data repositories](https://sinarueeger.github.io/post/the-headache-with-data-repositories/).

It's also remarkable that this started as an [unconf18 project](https://ropensci.org/blog/2018/06/05/unconf_recap_1/#dataspice) 🚀 and y'all have taken it this far.

Let me know what you think :-)
