### 2021-05-16 #426 @amoeba
Thank you both so much, @aebratt and @tdjames1, for your detailed reviews. I particularly appreciate the amount of time you invested and the level of detail at which you combed over the package. Would it be alright if we listed you both in the package description with the `rev` role code? If you have an ORCID, let me know what it is and I'll put it alongside your entries.

Hey @emilyriederer, thanks for handling this submission. I've gone ahead and transferred the package to https://github.com/ropensci and completed the rest of the changes on your checklist above. I think one or two final details might need taking care of but we should otherwise be good to go.

Re: a tech note or blog post, I have it in the back of my mind that I promised @stefaniebutland a blog post a while ago. I can't commit to one right now but I'll re-evaluate in a few weeks and hopefully I can carve out some time. I think a blog post is _very_ good idea.
