### 2021-05-16 #426 @tdjames1
Hi @amoeba, I'd be happy with being listed in the package description. My ORCID is 0000-0003-1363-4742. Thanks and good work getting the package approved!
